+++ client/src/components/global/Header/Header.js
@@ -8,6 +8,7 @@ import UI from '../../../ui';
 import * as storeUtils from '../../../storeUtils';
 import * as classes from '../../../constants/classes';
 import * as screens from '../../../constants/screens';
+import ReactSVG from 'react-svg';
 
 
 class Header extends React.Component {
@@ -42,7 +43,12 @@ class Header extends React.Component {
                 </NavLink>
               </div>
               <div className='Header__item__wrapper'>
-                <NavLink exact to={screens.HUB_PROJECT_EXPERIMENT_INDEX}>
+                <NavLink
+                  to={screens.HUB_PROJECT_EXPERIMENT_DEFAULT}
+                  className={classNames({
+                    active: window.location.pathname.startsWith('/experiment'),
+                  })}
+                >
                   <div className='Header__item'>
                     <UI.Icon i='nc-layers-3' className='Header__item__icon' />
                     <UI.Text className='Header__item__title'>Runs</UI.Text>
@@ -50,7 +56,12 @@ class Header extends React.Component {
                 </NavLink>
               </div>
               <div className='Header__item__wrapper'>
-                <NavLink to={screens.HUB_PROJECT_TAGS}>
+                <NavLink
+                  to={screens.HUB_PROJECT_TAGS}
+                  className={classNames({
+                    active: window.location.pathname.startsWith('/tag'),
+                  })}
+                >
                   <div className='Header__item'>
                     <UI.Icon i='nc-flag-points-32' className='Header__item__icon' />
                     <UI.Text className='Header__item__title'>Tags</UI.Text>
@@ -58,13 +69,36 @@ class Header extends React.Component {
                 </NavLink>
               </div>
               <div className='Header__item__wrapper'>
-                <NavLink to={screens.HUB_PROJECT_EXECUTABLES}>
+                <NavLink
+                  to={screens.HUB_PROJECT_EXECUTABLES}
+                  className={classNames({
+                    active: window.location.pathname.startsWith('/process'),
+                  })}
+                >
                   <div className='Header__item'>
                     <UI.Icon i='nc-archive-2' className='Header__item__icon' />
                     <UI.Text className='Header__item__title'>Processes</UI.Text>
                   </div>
                 </NavLink>
               </div>
+              {this.props.project.tf_enabled &&
+                <div className='Header__item__wrapper'>
+                  <NavLink
+                    to={screens.HUB_TF_SUMMARY_LIST}
+                    className={classNames({
+                      active: window.location.pathname.startsWith('/tf'),
+                    })}
+                  >
+                    <div className='Header__item'>
+                      <ReactSVG
+                        className='Header__item__icon__svg'
+                        src={require('../../../asset/icons/tensorflow-2.svg')}
+                      />
+                      <UI.Text className='Header__item__title'>TF logs</UI.Text>
+                    </div>
+                  </NavLink>
+                </div>
+              }
             </div>
             <div className='Header__items bottom'>
               <div className='Header__item__wrapper'>