+++ client/src/screens/hub/HubMainScreen/components/Panel/Panel.js
@@ -165,9 +165,11 @@ class Panel extends Component {
     }
 
     this.drawArea().then(() => this.drawAxes()).then(() => {
-      this.drawLines();
-      this.drawHoverAttributes();
-      this.bindInteractions();
+      window.requestAnimationFrame(() => {
+        this.drawLines();
+        this.drawHoverAttributes();
+        this.bindInteractions();
+      });
     });
   };