+++ client/src/screens/hub/HubTagCreateScreen/HubTagCreateScreen.js
@@ -89,7 +89,7 @@ class HubTagCreateScreen extends React.Component {
 
     return (
       <ProjectWrapper>
-        <UI.Container size='small' ref={this.contentRef}>
+        <UI.Container size='small'>
           {this._renderContent()}
         </UI.Container>
       </ProjectWrapper>