+++ client/src/screens/hub/HubTagsScreen/HubTagsScreen.js
@@ -102,7 +102,7 @@ class HubTagsScreen extends React.Component {
           <meta title='' content='' />
         </Helmet>
 
-        <UI.Container size='small' ref={this.contentRef}>
+        <UI.Container size='small'>
           {this._renderContent()}
         </UI.Container>
       </ProjectWrapper>