+++ client/src/utils/index.js
@@ -104,6 +104,18 @@ export function formatSize(size) {
   return [ceilSize, metric];
 }
 
+export function formatDuration(durationInSeconds) {
+  const hours = Math.floor(durationInSeconds / 3600);
+  const minutes = Math.floor((durationInSeconds % 3600) / 60);
+  const seconds = Math.floor((durationInSeconds % 3600) % 60);
+
+  return {
+    hours,
+    minutes,
+    seconds,
+  };
+}
+
 export function isDev() {
   return window.location.hostname === 'aim-dev.loc';
 }